added AnalogyEvaluator to tflab/embedding_utils.py
